pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.13",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.17.1",
    "gunicorn>=25.3.0",
    "python-dotenv>=1.2.2",
    "redis>=7.4.0",
    "requests>=2.33.1",
]
